Add register example that doesnt use the Identify protocol
The external address is specified manually.
